Move tests out of pages directory
We can't colocate these because everything in pages is assumed to be a
page by nextjs. This meant you could go to /index.test and it would try
to render the test. It also meant export and build were generating files
for the tests and snapshots.

Moved these to a pages_tests folder instead.